Add spp_type
Structure representing type is needed to be able to determine if the
object is of a given type. Also, having it in a separate struct allows
us to allocate and initalize a single object for each type, and refer to
it from every object implementing that type, as it is constant and does
not change ever.